+++ tick_utils.h
@@ -1,6 +1,6 @@
 /*!\file tick_utils.h
 ** \author SMFSW
-** \copyright MIT (c) 2017-2023, SMFSW
+** \copyright MIT (c) 2017-2024, SMFSW
 ** \brief Core ticks related utilities
 ** \warning Of course, these blocking delays are not designed to be used in an RTOS environment where they wouldn't work properly.
 ** \warning M0/M0+ cores doesn't have core debug trace peripheral, delay shall be generated using a dedicated TIM peripheral.
@@ -21,12 +21,19 @@
 /****************************************************************/
 
 
+/*!\brief Get number of ticks per microsecond (for reference)
+** \note Only relevant after call to \ref init_Delay_Generator
+** \return Number of tick per microsecond
+**/
+float get_TicksPerMicrosecond(void);
+
+
 /*!\brief Delay generator initialization function
 ** \note In case of system core clock frequency change during execution (calls to SystemCoreClockUpdate, HAL_RCC_ClockConfig), init_Delay_us shall be called to update tick ratio.
 ** \note Debug Trace peripheral timer only: On some Cortex cores, need to enable access to DWT prior to init and use of delays (automatically done only for M7 core).
-** \return HAL Status
+** \return FctERR - Error code
 **/
-HAL_StatusTypeDef init_Delay_Generator(void);
+FctERR init_Delay_Generator(void);
 
 /*!\brief Microseconds delay generator (blocking)
 ** \note Please keep in mind Delay_us is blocking code execution for the given amount of time (except interrupts).