+++ time_utils.c
@@ -1,6 +1,6 @@
 /*!\file time_utils.c
 ** \author SMFSW
-** \copyright MIT (c) 2017-2023, SMFSW
+** \copyright MIT (c) 2017-2024, SMFSW
 ** \brief Time related utilities
 **/
 /****************************************************************/