+++ ReleaseNotes.md
@@ -2,7 +2,7 @@
 
 The MIT License (MIT)
 
-Copyright (c) 2017-2023 SMFSW (Sebastien Bizien)
+Copyright (c) 2017-2024 SMFSW (Sebastien Bizien)
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -22,6 +22,9 @@ SOFTWARE.
 
 ## v1.6
 
+* Doxygen comments review and fixes
+* tick_utils: init function returns FctERR instead of HAL_StatusTypeDef
+* workload: Module added
 * PWM_IC: get_PWM_IC_LastUpdate added (useful to evaluate if IC is ongoing and resulting values aren't outdated)
 * More HAL_ENABLED checks
 * USB_CDC_IF_ex: USE_USB_CDC_IF_EX symbol check to enable features